Add TransactionRecordType to strenghen the type of deserialisation
The deserialisation currently uses a free text form for transaction
types. This is not ideal. This commit adds a new type named
`TransactionRecordType` to make the type stronger.

However, it would have been better if we could make the deserialisation
directly on top of the strongly typed
`transaction_processor::Transaction`:
```rust
/// The transaction structure accepted by this application.
pub struct Transaction {
    pub client_id: ClientId,
    pub transaction_id: TransactionId,
    pub kind: TransactionKind,
}

/// The kinds of transactions.
pub enum TransactionKind {
    DepositTransaction { amount: Amount },
    WithdrawalTransaction { amount: Amount },
    DisputeTransaction,
    ResolveTransaction,
    ChargeBackTransaction,
}
```
but it may not be trivial to do so. This effort is for now deferred to a
later stage.