Drop the use of mockall
The structs created by mockall is not so easy to harness especially when
you want to use the real struct in certain tests (e.g. integration
test).

It could still be that I was not using it properly but given the time
limit I have to drop the use of it, at least for now.