Add AsyncCsvStreamProcessor
This async implementation provides concurrent process of the input
stream. Although the input is still read sequentially, it offloads the
process of each read transaction to its worker queue by sending them
through a message channel, which is allocated one per client / account.

Such parallelism is at its maximum level since transactino records are
ordered as they arrive and a total order for each client / account is
required in order to maintain the correctness of the transactions. The
transactions of different clients / accounts do not have an order among
them and therefore can be processed concurrently.