Rename AccountTransactionProcessor to AccountTransactor
Apparently the term of Processor is too overloaded. Naming was done
in a rush now it's time to fix it. Transactor is a better name here
since it actually performs the transation, as opposed to its caller who
takes in a transaction record and pass it around to get it take effect.