Add guardrail for duplicate deposit and withdrawal
Although it is by requirement that the transaction id is globally
unique, it just feels safer that we know that this invariant is not
broken.

This commit adds an assert in there to ensure that if a duplicate
deposit or withdrawal is observed, it has to match the amount previously
seen.