Enhance idempotency all transactors
Currently, when an account is locked, even reperforming an operation
that has been performed previously would fail:
```
type,       client, tx, amount
deposit,         2,  2,    3.0
dispute,         2,  2,
chargeback,      2,  2,
deposit,         2,  2,    3.0  <-- this would fail
```
Generally, a locked account should reject all transaction (except the
one that unlocks it but that is out of the scope here). However, if the
process of a transaction stream fails due to some reason, and the
accounts have been updated up to the point of the stream at which the
failure occurred, it would be inconvenient to fix the problem as
reapplying the stream may not work if any account has been locked during
the process of the stream.

Making the transactors accept previously accepted operations as a
duplicate could ease this hassle.

However, this comes with a potential security issue that by issuing
operations to the transactor, one can peek into the transaction history.